Move buffer_s3response to shared
